refactor(ui): extract AutomationRepoFields from AutomationSettings, drop grandfather (#855)
Move the self-contained per-repo config-fields cluster (default-model override,
auto-drain numeric fields, repo roles) — none of which use the shared info/detail
snippets — plus its state, two $effects (drain re-seed, roles load) and helpers
into a child. Props: repoPath + autoDrain (= flags.autoDrain), both read-only.
AutomationSettings <template> 50/57 -> 39/37; child 12/20. Both clear the
Tier-1 40/60 bar, so the AutomationSettings .fallowrc.jsonc grandfather is
deleted.
